refactor: restructure home module to follow FSD conventions
- Move section components from components/ to features/
- Fix ContactModel.tsx typo to ContactModal.tsx
- Remove empty folders (hooks/, types/, ui/)
- Remove FaqSection template button (links to placeholder page)
- Remove TemplatesPage.tsx and its /templates route
- Delete FairWorklyHomePage-Test/ prototype files
- Update import paths and barrel exports
