+++ frontend/src/modules/home/features/FeaturesSection.tsx
