--- frontend/src/modules/home/pages/TemplatesPage.tsx
@@ -1,8 +1,0 @@
-export function TemplatesPage() {
-  return (
-    <div>
-      <h1>TemplatesPage</h1>
-      <p>Coming soon...</p>
-    </div>
-  )
-}
